Added protections for when SDL TTF library doesnt exist.
Instead of seg-faulting, we just don't display any text. which is
better.
Later on, we probably want to integrate that functionality into the text
class.
Maybe we even want to move loading SDL TTF into the text class too?
But that would require some way of checking to see if it is loaded
already, which Derelict doesnt provide.